moved getTimestamp() from logger -> utility
