+++ includes/Utility.hpp
@@ -6,4 +6,5 @@ namespace Util
 {
 	std::string removeTrailingWhitespace(const std::string &str);
 	std::string removeLeadingWhitespace(const std::string &str);
+	std::string getTimeStamp();
 }
