+++ srcs/Logger.cpp
@@ -1,27 +1,12 @@
-#include <Logger.hpp>
 #include <Colors.hpp>
+#include <Logger.hpp>
+#include <Utility.hpp>
 
 #include <ctime>
 #include <sstream>
 
 std::ofstream Logger::_logFile(LOGFILE);
 std::stringstream Logger::inputStream;
-
-static std::string getTimeStamp()
-{
-	std::string timestamp;
-	struct tm* timeinfo;
-	time_t rawtime;
-
-	time(&rawtime);
-	timeinfo = localtime(&rawtime);
-
-	timestamp = "[";
-	timestamp += asctime(timeinfo);
-	timestamp.pop_back();
-	timestamp += "] ";
-	return timestamp;
-}
 
 static void printError(const std::string& msg)
 {
@@ -38,9 +23,6 @@ static void printDebug(const std::string& msg)
 	std::cout << C_CYAN << "[DEBUG] " << C_RESET << msg << std::endl;
 }
 
-
-
-
 void Logger::log(const std::string& msg)
 {
 	if (!ENABLE_LOGGING)
@@ -52,7 +34,7 @@ void Logger::log(const std::string& msg)
 		return;
 	}
 
-	std::string timestamp = getTimeStamp();
+	std::string timestamp = Util::getTimeStamp();
 
 	_logFile.write(timestamp.c_str(), timestamp.size());
 	_logFile.write(msg.c_str(), msg.size());
